+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.1-preview.1] - 2021-07-12
+### Added
+- Camera-based realtime reflection
+
+### Fixed
+- WebGL runtime error for Unity2020 build due to unsupported compression setting.
+
 
 ## [1.1.0-preview.2] - 2021-07-05
 ### Added